requirements.txt
requests
websocket-client
aiohttp
python-socketio